+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidyBdE (development version)
 
+-   Add vignette to package
+
 # tidyBdE 0.1.0
 
 -   Initial release.