+++ NEWS.md
@@ -1,6 +1,7 @@
 # tidyBdE (development version)
 
 -   Add vignette to package
+-   Color adjustment on `bde_rose_pal()`: Now hcl spectrum is more consistent between roses and blues according to `colorspace::specplot()`.
 
 # tidyBdE 0.1.0