+++ NEWS.md
@@ -2,6 +2,7 @@
 
 -   Add vignette to package
 -   Color adjustment on `bde_rose_pal()`: Now hcl spectrum is more consistent between roses and blues according to `colorspace::specplot()`.
+-   Add DOI: https://doi.org/10.5281/zenodo.4673496
 
 # tidyBdE 0.1.0